feature/new-cli-manifest
