+++ pipelex/cli/agent_cli/commands/inputs/pipe_cmd.py
@@ -1,7 +1,8 @@
-"""Agent CLI inputs command - generate example inputs for a pipe with JSON output."""
+"""Agent CLI inputs pipe command - generate example inputs by pipe code or bundle path."""
+
+from __future__ import annotations
 
 import asyncio
-import json
 from pathlib import Path
 from typing import Annotated, Any
 
@@ -9,74 +10,16 @@ import typer
 
 from pipelex.cli.agent_cli.commands.agent_cli_factory import make_pipelex_for_agent_cli
 from pipelex.cli.agent_cli.commands.agent_output import agent_error, agent_success, extract_validation_errors
+from pipelex.cli.agent_cli.commands.inputs._inputs_core import inputs_core
 from pipelex.core.interpreter.helpers import is_pipelex_file
 from pipelex.core.pipes.exceptions import PipeOperatorModelChoiceError
 from pipelex.core.pipes.inputs.exceptions import NoInputsRequiredError
-from pipelex.core.pipes.inputs.input_renderer import render_inputs
-from pipelex.hub import (
-    get_library_manager,
-    get_required_pipe,
-    resolve_library_dirs,
-    set_current_library,
-)
 from pipelex.pipe_operators.exceptions import PipeOperatorModelAvailabilityError
 from pipelex.pipelex import Pipelex
-from pipelex.pipeline.validate_bundle import ValidateBundleError, validate_bundle
+from pipelex.pipeline.validate_bundle import ValidateBundleError
 
 
-async def _inputs_core(
-    pipe_code: str | None = None,
-    bundle_path: Path | None = None,
-    library_dirs: list[Path] | None = None,
-) -> dict[str, Any]:
-    """Core logic for generating input JSON for a pipe.
-
-    Args:
-        pipe_code: The pipe code to generate inputs for.
-        bundle_path: Path to the bundle file (.mthds).
-        library_dirs: List of library directories to search for pipe definitions.
-
-    Returns:
-        Dictionary with inputs suitable for JSON serialization.
-
-    Raises:
-        ValidateBundleError: If bundle validation fails.
-        NoInputsRequiredError: If the pipe has no inputs.
-    """
-    if bundle_path:
-        validate_bundle_result = await validate_bundle(mthds_file_path=bundle_path, library_dirs=library_dirs)
-        bundle_blueprint = validate_bundle_result.blueprints[0]
-        if not pipe_code:
-            main_pipe_code = bundle_blueprint.main_pipe
-            if not main_pipe_code:
-                msg = f"Bundle '{bundle_path}' does not declare a main_pipe. Specify a pipe code with --pipe."
-                raise ValidateBundleError(message=msg)
-            pipe_code = main_pipe_code
-    else:
-        # No bundle - initialize the library manually
-        library_manager = get_library_manager()
-        library_id, _ = library_manager.open_library()
-        set_current_library(library_id=library_id)
-        effective_dirs, _ = resolve_library_dirs(library_dirs)
-        if effective_dirs:
-            library_manager.load_libraries(library_id=library_id, library_dirs=effective_dirs)
-
-    if not pipe_code:
-        msg = "No pipe code specified"
-        raise ValidateBundleError(message=msg)
-
-    the_pipe = get_required_pipe(pipe_code=pipe_code)
-    inputs_json_str = render_inputs(the_pipe, indent=2)
-    inputs_dict = json.loads(inputs_json_str)
-
-    return {
-        "success": True,
-        "pipe_code": pipe_code,
-        "inputs": inputs_dict,
-    }
-
-
-def inputs_cmd(
+def inputs_pipe_cmd(
     ctx: typer.Context,
     target: Annotated[
         str | None,
@@ -96,10 +39,10 @@ def inputs_cmd(
     Outputs JSON to stdout on success, JSON to stderr on error with exit code 1.
 
     Examples:
-        pipelex-agent inputs my_pipe
-        pipelex-agent inputs my_bundle.mthds
-        pipelex-agent inputs my_bundle.mthds --pipe my_pipe
-        pipelex-agent inputs my_pipe -L ./my_pipes
+        pipelex-agent inputs pipe my_pipe
+        pipelex-agent inputs pipe my_bundle.mthds
+        pipelex-agent inputs pipe my_bundle.mthds --pipe my_pipe
+        pipelex-agent inputs pipe my_pipe -L ./my_pipes
     """
     # Validate that at least one target is provided
     if target is None and pipe is None:
@@ -134,7 +77,7 @@ def inputs_cmd(
     make_pipelex_for_agent_cli(library_dirs=library_dirs, log_level=ctx.obj["log_level"])
 
     try:
-        result = asyncio.run(_inputs_core(pipe_code=pipe_code, bundle_path=bundle_path, library_dirs=library_dirs))
+        result = asyncio.run(inputs_core(pipe_code=pipe_code, bundle_path=bundle_path, library_dirs=library_dirs))
         agent_success(result)
 
     except FileNotFoundError as exc:
