+++ pipelex/cli/agent_cli/commands/run/pipe_cmd.py
@@ -1,4 +1,6 @@
-"""Agent CLI run command - simplified pipeline execution with JSON output."""
+"""Agent CLI run pipe command - execute a pipeline by pipe code or bundle path."""
+
+from __future__ import annotations
 
 import asyncio
 import json
@@ -10,142 +12,19 @@ import typer
 from pipelex.builder.conventions import DEFAULT_BUNDLE_FILE_NAME, DEFAULT_INPUTS_FILE_NAME
 from pipelex.cli.agent_cli.commands.agent_cli_factory import make_pipelex_for_agent_cli
 from pipelex.cli.agent_cli.commands.agent_output import agent_error, agent_success
-from pipelex.config import get_config
+from pipelex.cli.agent_cli.commands.run._run_core import run_pipeline_core
+from pipelex.cli.method_resolver import resolve_pipe_from_exports
 from pipelex.core.interpreter.exceptions import MthdsDecodeError, PipelexInterpreterError
 from pipelex.core.interpreter.helpers import MTHDS_EXTENSION, is_pipelex_file
 from pipelex.core.interpreter.interpreter import PipelexInterpreter
 from pipelex.core.pipes.exceptions import PipeOperatorModelChoiceError
-from pipelex.graph.graph_factory import generate_graph_outputs, save_graph_outputs_to_dir
 from pipelex.pipe_operators.exceptions import PipeOperatorModelAvailabilityError
-from pipelex.pipe_run.pipe_run_mode import PipeRunMode
 from pipelex.pipelex import Pipelex
 from pipelex.pipeline.exceptions import PipelineExecutionError
-from pipelex.pipeline.runner import PipelexRunner
-from pipelex.tools.misc.json_utils import JsonTypeError, clean_json_dumps, load_json_dict_from_path
+from pipelex.tools.misc.json_utils import JsonTypeError, load_json_dict_from_path
 
 
-async def _run_pipeline_core(
-    pipe_code: str,
-    mthds_content: str | None = None,
-    bundle_uri: str | None = None,
-    inputs: dict[str, Any] | None = None,
-    dry_run: bool = False,
-    mock_inputs: bool = False,
-    library_dirs: list[str] | None = None,
-    graph: bool = False,
-) -> dict[str, Any]:
-    """Core logic for running a pipeline and returning JSON-serializable output.
-
-    Args:
-        pipe_code: The pipe code to run.
-        mthds_content: MTHDS content string (optional).
-        bundle_uri: Bundle file path (optional).
-        inputs: Input dictionary for the pipeline.
-        dry_run: Whether to run in dry mode (no actual inference).
-        mock_inputs: Whether to generate mock data for missing inputs.
-        library_dirs: List of library directories to search for pipe definitions.
-        graph: Whether to generate execution graph visualizations.
-
-    Returns:
-        Dictionary with execution results suitable for JSON serialization.
-
-    Raises:
-        PipelineExecutionError: If the pipeline execution fails.
-    """
-    pipe_run_mode = PipeRunMode.DRY if dry_run else None
-
-    execution_config = get_config().pipelex.pipeline_execution_config.with_graph_config_overrides(
-        generate_graph=graph,
-        mock_inputs=mock_inputs or None,
-    )
-
-    runner = PipelexRunner(
-        bundle_uri=bundle_uri,
-        pipe_run_mode=pipe_run_mode,
-        execution_config=execution_config,
-        library_dirs=library_dirs,
-    )
-    response = await runner.execute_pipeline(
-        pipe_code=pipe_code,
-        mthds_content=mthds_content,
-        inputs=inputs,
-    )
-    pipe_output = response.pipe_output
-
-    main_stuff = pipe_output.working_memory.get_optional_main_stuff()
-    main_stuff_json: dict[str, Any] = {}
-    if main_stuff:
-        main_stuff_json = {
-            "json": await main_stuff.content.rendered_json_async(),
-            "markdown": await main_stuff.content.rendered_markdown_async(),
-            "html": await main_stuff.content.rendered_html_async(),
-        }
-
-    result: dict[str, Any] = {
-        "success": True,
-        "pipe_code": pipe_code,
-        "dry_run": dry_run,
-        "main_stuff": main_stuff_json,
-        "working_memory": pipe_output.working_memory.smart_dump(),
-    }
-
-    # Determine output directory: next to the bundle, or pipelex-wip/ fallback
-    output_dir: Path
-    if bundle_uri:
-        output_dir = Path(bundle_uri).parent
-    else:
-        output_dir = Path("pipelex-wip")
-    output_dir.mkdir(parents=True, exist_ok=True)
-
-    # Generate and save graph visualizations if requested
-    if graph and pipe_output.graph_spec:
-        graph_config = execution_config.graph_config
-        # Enable ReactFlow HTML output and data inclusion for the render
-        render_graph_config = graph_config.model_copy(
-            update={
-                "data_inclusion": graph_config.data_inclusion.model_copy(
-                    update={
-                        "stuff_json_content": True,
-                        "stuff_text_content": True,
-                        "stuff_html_content": True,
-                    }
-                ),
-                "graphs_inclusion": graph_config.graphs_inclusion.model_copy(
-                    update={
-                        "graphspec_json": False,
-                        "mermaidflow_html": False,
-                        "reactflow_html": True,
-                    }
-                ),
-            }
-        )
-
-        graph_outputs = await generate_graph_outputs(
-            graph_spec=pipe_output.graph_spec,
-            graph_config=render_graph_config,
-            pipe_code=pipe_code,
-        )
-
-        saved_files = save_graph_outputs_to_dir(graph_outputs=graph_outputs, output_dir=output_dir)
-
-        # Rename reactflow.html to mode-aware filename
-        graph_filename = "dry_run.html" if dry_run else "live_run.html"
-        reactflow_path = saved_files.get("reactflow_html")
-        if reactflow_path:
-            final_path = reactflow_path.parent / graph_filename
-            reactflow_path.rename(final_path)
-            result["graph_files"] = {"graph_html": str(final_path)}
-
-    # Save output JSON after all result fields are populated
-    output_filename = "dry_run.json" if dry_run else "live_run.json"
-    output_path = output_dir / output_filename
-    output_path.write_text(clean_json_dumps(result, indent=2), encoding="utf-8")
-    result["output_file"] = str(output_path)
-
-    return result
-
-
-def run_cmd(
+def run_pipe_cmd(
     ctx: typer.Context,
     target: Annotated[
         str | None,
@@ -180,16 +59,16 @@ def run_cmd(
         typer.Option("--library-dir", "-L", help="Directory to search for pipe definitions (.mthds files)"),
     ] = None,
 ) -> None:
-    """Execute a pipeline and output JSON results.
+    """Execute a pipeline by pipe code or bundle path and output JSON results.
 
     Outputs JSON to stdout on success, JSON to stderr on error with exit code 1.
     Graph visualizations are generated by default; use --no-graph to disable.
 
     Examples:
-        pipelex-agent run pipelex-wip/pipeline_01/
-        pipelex-agent run my_bundle.mthds --pipe my_pipe
-        pipelex-agent run my_pipe --dry-run --mock-inputs
-        pipelex-agent run pipelex-wip/pipeline_01/ --no-graph
+        pipelex-agent run pipe pipelex-wip/pipeline_01/
+        pipelex-agent run pipe my_bundle.mthds --pipe my_pipe
+        pipelex-agent run pipe my_pipe --dry-run --mock-inputs
+        pipelex-agent run pipe pipelex-wip/pipeline_01/ --no-graph
     """
     # Validate that at least one target is provided
     provided_options = sum([target is not None, pipe is not None, bundle is not None])
@@ -263,6 +142,15 @@ def run_cmd(
     if not pipe_code and not bundle_path:
         agent_error("No pipe code or bundle file specified", "ArgumentError")
 
+    # Check installed methods' exports for additional library dirs
+    if pipe_code and not bundle_path:
+        export_dirs = resolve_pipe_from_exports(pipe_code)
+        if export_dirs:
+            if library_dir is None:
+                library_dir = export_dirs
+            else:
+                library_dir = [*export_dirs, *library_dir]
+
     # Load MTHDS content from bundle if provided
     mthds_content: str | None = None
     if bundle_path:
@@ -304,7 +192,7 @@ def run_cmd(
 
     try:
         result = asyncio.run(
-            _run_pipeline_core(
+            run_pipeline_core(
                 pipe_code=pipe_code,  # type: ignore[arg-type]
                 mthds_content=mthds_content,
                 bundle_uri=bundle_path,
