+++ tests/e2e/pipelex/cli/test_validate_cmd.py
@@ -1,6 +1,6 @@
 from pathlib import Path
 
-from pipelex.cli.commands.validate_cmd import do_validate_all_libraries_and_dry_run
+from pipelex.cli.commands.validate._validate_core import do_validate_all_libraries_and_dry_run  # noqa: PLC2701
 
 
 class TestValidateCommand:
